pyproject.toml
[tool.poetry]
name = "modular_bna"
version = "0.1.0"
description = ""
authors = ["Rémy Greinhofer <remy.greinhofer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
better-exceptions = "^0.3.3"
brokenspoke-analyzer = {git = "https://github.com/PeopleForBikes/brokenspoke-analyzer", rev = "main"}
typer = "^0.9.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
rich = "^13.4.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipykernel = "^6.23.2"
isort = "^5.12.0"
pandas = "^2.0.2"
pytest-asyncio = "^0.21.0"
pytest-reporter-html1 = "^0.8.3"
pytest-rerunfailures = "^11.1.2"
ruff = "^0.0.270"
sqlfluff = "^2.1.1"
xdoctest = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.sqlfluff.core]
dialect = "postgres"
large_file_skip_byte_limit = 40000

[tool.isort]
profile = "black"
force_grid_wrap = 2

[tool.pytest.ini_options]
minversion = "6.0"
addopts = """
  -p no:warnings \
  --reruns 4 --reruns-delay 15 --rerun-except assert \
  --template=html1/index.html --report=report.html \
"""
markers = [
    "australia",
    "canada",
    "europe",
    "france",
    "spain",
    "usa",
    "xs: runs under 5min",
    "s: runs under 15min",
    "m: runs under 60min (1h)",
    "l: runs under 180min (2h)",
    "xl: runs under 360min (6h)",
    "xxl: runs under 720min (12h / 1/2day)",
]

[tool.poetry.scripts]
modbna = 'modular_bna.cli:app'
